chore: merged into structured mod
