+++ promkit-widgets/src/structured/json/document.rs
